Added sword beam but currently its just an arrow sprite I got lazy. bomb is implemented to explode and does a small animation of getting bigger.

+++ Sprint2/ItemClasses/IItem.cs
@@ -9,7 +9,7 @@ namespace Sprint2
     interface IItem
     {
         public void SetSpriteActivity(Boolean activity);
-        public void SetSize(int size);
+        public void SetSize(double size);
         public void SetPosition(Point position);
         public Boolean SpriteActive();
         public void Update(GameTime gameTime);
