+++ Sprint2/SpriteClasses/EnemySprite/SkeletonSprite.cs
@@ -30,7 +30,7 @@ namespace Sprint2
                
         }
 
-        public void SetSize(int size)
+        public void SetSize(double size)
         {
 
             //nothing?? for now....
